Make the quiz totally random. Also don't time out mid-question
This feels waaay better. Also the timer ending before having a chance to type
in the answer was surprisingly infuriating.

Test Plan:
Spend a bunch of time playing with it and notice that I'm actually having to
recall (or think about) each fact, rather than feeling like it's more
pattern-matchy